+++ tests/testthat/helper.R
@@ -7,3 +7,17 @@ get_os_string <- function(unix_string, windows_string) {
 
     output
 }
+
+find_string <- function(value, lines = c()) {
+    found <- FALSE
+
+    for (line in lines) {
+        found <- grepl(value, line)
+
+        if (found) {
+            break
+        }
+    }
+
+    found
+}