+++ tests/testthat/test_generate_env_setup_script.R
@@ -6,8 +6,7 @@ describe("generate_env_setup_script", {
     it("no env", {
         script <- scriptexec::generate_env_setup_script()
 
-        length <- nchar(script)
-
+        length <- length(script)
         expect_equal(length, 0)
     })
 
@@ -15,15 +14,12 @@ describe("generate_env_setup_script", {
         script <- scriptexec::generate_env_setup_script(c("ENV1=VALUE1",
             "ENV2=VALUE2"))
 
-        found <- grepl("ENV1=VALUE1", script)
-        expect_true(found, TRUE)
-
-        found <- grepl("ENV2=VALUE2", script)
-        expect_true(found, TRUE)
+        length <- length(script)
+        expect_equal(length, 2)
 
         prefix <- get_os_string("export", "SET")
-        expected_result <- paste(paste(prefix, "ENV1=VALUE1", sep = " "),
-            paste(prefix, "ENV2=VALUE2", sep = " "), sep = "\n")
+        expected_result <- c(paste(prefix, "ENV1=VALUE1", sep = " "), paste(prefix,
+            "ENV2=VALUE2", sep = " "))
 
         expect_equal(expected_result, script)
     })